+++ src/components/form/field/DatetimeField.tsx
@@ -1,13 +1,14 @@
 import { FieldValues } from 'react-hook-form'
 import { InputFieldProps } from './define'
 
-import DatetimeInput from '../base/DatetimeInput'
+import DateInput from '../base/DateInput'
 
 export default function DatetimeField<T extends FieldValues>(
   props: InputFieldProps<'datetime', T>,
 ) {
   return (
-    <DatetimeInput
+    <DateInput
+      includeTime={true}
       className={props.formInput.coreClassName}
       {...props.formInput.attributes}
       {...props.useFormReturns.register(
